Deprecate tests for `V3DBHandler`
